Post hash tag image

+++ Backend/blog/blog/src/main/java/com/main/blog/service/BlockService.java
@@ -14,6 +14,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Set;
+import java.util.stream.Collectors;
+
 @Service
 public class BlockService implements IBlockService {
 
@@ -31,28 +34,41 @@ public class BlockService implements IBlockService {
 
     @Override
     public ResponseEntity<?> getUserBans(Long userId) {
-        return ResponseEntity.status(HttpStatus.OK).body(iBlockRepository.findByBlockedUserId(userId));
+        Set<BlockDto> blocks = iBlockRepository.findByBlockingUserId(userId).stream().map(block -> mapper.getMapper().map(block, BlockDto.class)).collect(Collectors.toSet());
+        return ResponseEntity.status(HttpStatus.OK).body(blocks);
     }
 
     @Override
     public ResponseEntity<?> addBlock(Long userId, String token) {
-        User userToBlock = iUserRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("User to block not found"));
         try {
+            if (token.startsWith("Bearer ")) {
+                token = token.substring(7);
+            }
+            User userToBlock = iUserRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("User to block not found"));
             User userBlocking = iUserRepository.findByEmail(jwtService.extractUserEmail(token));
             if (userBlocking == null) {
                 throw new ResourceNotFoundException("User not found");
             }
             Block block = Block.builder().blockedUser(userToBlock).blockingUser(userBlocking).build();
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body(mapper.getMapper().map(block, Block.class));
+            iBlockRepository.save(block);
+            return ResponseEntity.status(HttpStatus.ACCEPTED).body(mapper.getMapper().map(block, BlockDto.class));
         } catch (ResourceNotFoundException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
         }
     }
 
     @Override
-    public ResponseEntity<?> deleteBlock(Long blockId) {
+    public ResponseEntity<?> deleteBlock(Long userId, String token) {
         try {
-            Block block = iBlockRepository.findById(blockId).orElseThrow(() -> new ResourceNotFoundException("Not found"));
+            if (token.startsWith("Bearer ")) {
+                token = token.substring(7);
+            }
+            User userBlocked = iUserRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("User to block not found"));
+            User userBlocking = iUserRepository.findByEmail(jwtService.extractUserEmail(token));
+            if (userBlocking == null) {
+                throw new ResourceNotFoundException("User not found");
+            }
+            Block block = iBlockRepository.findByBlockedUserIdAndBlockingUserId(userBlocked.getId(), userBlocking.getId());
             iBlockRepository.delete(block);
             return ResponseEntity.status(HttpStatus.ACCEPTED).body("User unlocked");
         } catch (ResourceNotFoundException e) {
